+++ Data Stream Host/monitormate.sh
@@ -1,13 +1,13 @@
 #!/bin/sh
 
 ### BEGIN INIT INFO
-# Provides:          monitormate
-# Required-Start:    $remote_fs $syslog
-# Required-Stop:     $remote_fs $syslog
-# Default-Start:     2 3 4 5
-# Default-Stop:      0 1 6
+# Provides:			 monitormate
+# Required-Start:	 $remote_fs $syslog
+# Required-Stop:	 $remote_fs $syslog
+# Default-Start:	 2 3 4 5
+# Default-Stop:		 0 1 6
 # Short-Description: Monitor the datastream from an Outback Mate3
-# Description:       Monitor the datastream from an Outback Mate3
+# Description:		 Monitor the datastream from an Outback Mate3
 ### END INIT INFO
 
 # Change the next 3 lines to suit where you install your script and what you want to call it
@@ -26,34 +26,34 @@ PIDFILE=/var/run/$DAEMON_NAME.pid
 . /lib/lsb/init-functions
 
 do_start () {
-    log_daemon_msg "Starting system $DAEMON_NAME daemon"
-    start-stop-daemon --start --background --pidfile $PIDFILE --make-pidfile --user $DAEMON_USER --startas $DAEMON -- $DAEMON_ARG
-    log_end_msg $?
+	log_daemon_msg "Starting system $DAEMON_NAME daemon"
+	start-stop-daemon --start --background --pidfile $PIDFILE --make-pidfile --user $DAEMON_USER --startas $DAEMON -- $DAEMON_ARG
+	log_end_msg $?
 }
 do_stop () {
-    log_daemon_msg "Stopping system $DAEMON_NAME daemon"
-    start-stop-daemon --stop --pidfile $PIDFILE --retry 10
-    log_end_msg $?
+	log_daemon_msg "Stopping system $DAEMON_NAME daemon"
+	start-stop-daemon --stop --pidfile $PIDFILE --retry 10
+	log_end_msg $?
 }
 
 case "$1" in
 
-    start|stop)
-        do_${1}
-        ;;
+	start|stop)
+		do_${1}
+		;;
 
-    restart|reload|force-reload)
-        do_stop
-        do_start
-        ;;
+	restart|reload|force-reload)
+		do_stop
+		do_start
+		;;
 
-    status)
-        status_of_proc "$DAEMON_NAME" "$DAEMON" && exit 0 || exit $?
-        ;;
-    *)
-        echo "Usage: /etc/init.d/$DEAMON_NAME {start|stop|restart|status}"
-        exit 1
-        ;;
+	status)
+		status_of_proc "$DAEMON_NAME" "$DAEMON" && exit 0 || exit $?
+		;;
+	*)
+		echo "Usage: /etc/init.d/$DEAMON_NAME {start|stop|restart|status}"
+		exit 1
+		;;
 
 esac
 exit 0
