+++ Data Stream Host/radian_class.py
@@ -1,22 +1,24 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
-# Filename: radian.py
+# Filename: radian_class.py
 
-#Copyright (C) 2012 Jesus Perez <jepefe@gmail.com>
-#This program is free software: you can redistribute it and/or modify
-#it under the terms of the GNU General Public License as published by
-#the Free Software Foundation, either version 2 of the License, or
-#(at your option) any later version.
+# Copyright (C) 2012-2014 Jesus Perez, Timothy Martin
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 2 of the License, or
+# (at your option) any later version.
 # 
-#This program is distributed in the hope that it will be useful,
-#but WITHOUT ANY WARRANTY; without even the implied warranty of
-#MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#GNU General Public License at <http://www.gnu.org/licenses/>
-#for more details.
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License at <http://www.gnu.org/licenses/>
+# for more details.
 
 
-
-class fxinv():
+#-----------------------------------#
+# Radian Inverter/Charger class		#
+#-----------------------------------#
+class radian():
     
     #Names of all data values in raw_string
     valuenames = ['address',\
